test: run PXI trajectory assertion in normal CI via vitest
Move the deterministic trajectory unit coverage from a playwright
.unit.spec.ts (which only ran under PXI_E2E=true) to app/__tests__/,
which vitest picks up in every CI run. The helper itself is pure logic
with no playwright dependency.
